pyproject.toml
[project]
name = "ragoveraudio"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.44.3",
    "generativeai>=0.0.1",
    "google>=3.0.0",
    "google-genai>=1.38.0",
    "langchain-text-splitters>=0.3.11",
    "numpy>=2.3.3",
    "pip>=25.2",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "voyageai>=0.3.5",
]
